requirements.txt
datetime
gradio
keras
matplotlib
mgzip
numpy
pyguitarpro
tensorflow
tqdm